lots of shit
I tried to do my best to tinker with the defines in small ways to improve it, unsure if this will be most effective for preventing AI stupidity.

Other than that, bugfixes, no more double napoleons, more balanced italian revolution and war with Austria, Austria slightly less likely to explode and die now.